NagrikAurSamvidhan-Backend/AIGeneration/requirements.txt
fastapi
transformers
langchain
langchain_groq
langchain_community
langchain_google_genai
python-dotenv
pypdf
faiss-cpu
uvicorn
unstructured
google-generativeai
pymongo
bson